feat: apply admin switch toggle at landing header
